Added Excel Create/Write/Read and Fleshed Template
Have added a custom library for creating , writing, and reading excel (ver. 2003) to the character template.

With this the character template java file, when run,  will take the info from the spread sheet, assign it to the character variables, and the print it out to show you.

Open the charactertemplate.java and hit run to see this work.

The file it is reading is the characterdata.xls. This is not a finished first version. More fleshing out is needed.